refactor: use `chains` config instead of `wagmi` config
